AP-2599: Refactor Interface services
Create a base class with shared code and then inherit it in
the Result and Submission services

This should reduce 2 more CodeClimate complaints around
duplicate code
