+++ app/services/hmrc/interface/submission_service.rb
@@ -1,15 +1,6 @@
 module HMRC
   module Interface
-    class SubmissionService
-      def self.call(hmrc_response)
-        new(hmrc_response).call
-      end
-
-      def initialize(hmrc_response)
-        @application = hmrc_response.legal_aid_application
-        @use_case = hmrc_response.use_case
-      end
-
+    class SubmissionService < BaseService
       def call
         raw_response = post_request
         parsed_response = parse_json_response(raw_response.body)
@@ -31,16 +22,20 @@ module HMRC
 
       private
 
+      def application
+        @application ||= @hmrc_response.legal_aid_application
+      end
+
+      def use_case
+        @use_case ||= @hmrc_response.use_case
+      end
+
       def applicant_values
-        @applicant_values ||= @application.applicant.json_for_hmrc
+        @applicant_values ||= application.applicant.json_for_hmrc
       end
 
       def date_values
-        @date_values ||= { start_date: @application.transaction_period_start_on, end_date: @application.transaction_period_finish_on }
-      end
-
-      def conn
-        @conn ||= Faraday.new(url: host, headers: headers)
+        @date_values ||= { start_date: application.transaction_period_start_on, end_date: application.transaction_period_finish_on }
       end
 
       def post_request
@@ -53,52 +48,8 @@ module HMRC
         catch_and_record_exception(e)
       end
 
-      def bearer_token
-        @bearer_token ||= OmniAuth::HMRC::Client.new.bearer_token
-      end
-
       def url_path
-        @url_path ||= "api/v1/submission/create/#{@use_case}"
-      end
-
-      def host
-        Rails.configuration.x.hmrc_interface.host
-      end
-
-      def headers
-        {
-          'Content-Type' => 'application/json',
-          'Accept' => 'application/json',
-          'Authorization' => "Bearer #{bearer_token}"
-        }
-      end
-
-      def catch_and_record_exception(error, http_method = 'POST')
-        raise_exception_error(
-          message: formatted_error_message(error),
-          backtrace: error.backtrace&.join("\n"),
-          http_method: http_method,
-          http_status: error.respond_to?(:http_status) ? error.http_status : nil
-        )
-      end
-
-      def raise_exception_error(message:, backtrace: nil, http_method: 'POST', http_status: nil)
-        Rails.logger.info { { message: message, backtrace: backtrace, method: http_method, http_status: http_status } }
-        raise HMRC::SubmissionError.new(message, http_status)
-      end
-
-      def formatted_error_message(err)
-        "#{self.class} received #{err.class}: #{err.message}"
-      end
-
-      def parse_json_response(response_body)
-        JSON.parse(response_body, symbolize_names: true)
-      rescue JSON::ParserError, TypeError
-        response_body || ''
-      end
-
-      def detailed_error(parsed_response, url)
-        "Bad Request: URL: #{url}, details: #{parsed_response}"
+        @url_path ||= "api/v1/submission/create/#{use_case}"
       end
     end
   end
